created lib and separated modules
