+++ iwp/src/img.rs
@@ -1,0 +1,46 @@
+use image::Luma;
+
+#[derive(Clone, Copy)]
+pub struct PixelT {
+    pub coords: (u32, u32),
+    pub value: u8,
+}
+
+pub enum ConnTypes {
+    Four = 4,
+    Eight = 8,
+}
+
+pub fn get_pixel_neighbours(
+    img: &image::ImageBuffer<Luma<u8>, Vec<u8>>,
+    coords: (u32, u32),
+    conn: ConnTypes,
+) -> Vec<(u32, u32)> {
+    let x = coords.0;
+    let y = coords.1;
+    let mut neighbours = Vec::new();
+
+    let floor_x = if x != 0 { x - 1 } else { x };
+    let floor_y = if y != 0 { y - 1 } else { y };
+
+    for i in (floor_x)..(x + 2) {
+        for j in (floor_y)..(y + 2) {
+            if !(i == x && j == y) {
+                match img.get_pixel_checked(i, j) {
+                    Some(_) => match conn {
+                        ConnTypes::Four => {
+                            if i == x || j == y {
+                                neighbours.push((i, j));
+                            }
+                        }
+
+                        ConnTypes::Eight => neighbours.push((i, j)),
+                    },
+                    None => continue,
+                }
+            }
+        }
+    }
+
+    return neighbours;
+}
